requirements.txt
addict==2.2.1
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.3
django-cors-headers==3.0.2
djangorestframework==3.10.1
djangorestframework-camel-case==1.0.3
drf-yasg==1.16.1
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
Markdown==3.1.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
pytz==2019.1
requests==2.22.0
ruamel.yaml==0.15.100
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
